requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
bcrypt==4.1.2
authlib==1.3.0
requests==2.31.0
